+++ backend/src/back_api/del_record.py
@@ -7,9 +7,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from back_api.rate_limiter import limiter
 from back_api.schemas import Message, RemoveRecordRequest
 from back_api.token import get_current_user
-from database.crud import remove_record_row_by_id
 from database.database import get_session
 from database.hash import decrypt_id
+from repository.record import remove_record_row_by_id
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
