+++ backend/src/repository/publication.py
@@ -1,0 +1,50 @@
+import logging
+from collections.abc import Sequence
+
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.future import select
+
+from database.models import Publ, Record
+
+logger = logging.getLogger(__name__)
+
+
+async def get_publication(session: AsyncSession, publ_id: int) -> Publ | None:
+    stmt = select(Publ).where(Publ.id == publ_id)
+    result = await session.execute(stmt)
+    return result.scalar_one_or_none()
+
+
+async def get_publications_for_language(
+    session: AsyncSession, language: str
+) -> Sequence[int]:
+    filters = [Publ.ural.is_(True), Publ.coords.is_(True), Publ.year > 1950]
+    if language != "all":
+        filters.append(Publ.language.ilike(f"%{language}%"))
+
+    stmt = select(Publ.id).where(*filters)
+    result = await session.execute(stmt)
+    return result.scalars().all()
+
+
+async def is_publ_filled(session: AsyncSession, user_id: int, publ_id: int) -> bool:
+    stmt = (
+        select(Record.type)
+        .where(Record.user_id == user_id, Record.publ_id == publ_id)
+        .order_by(Record.datetime.desc())
+        .limit(1)
+    )
+
+    result = await session.execute(stmt)
+    record_type = result.scalar_one_or_none()
+
+    if record_type is None:
+        return False
+
+    return record_type == "rec_ok"
+
+
+async def add_publication_from_json(session: AsyncSession, publ_json: dict) -> None:
+    publ = Publ(**publ_json)
+    session.add(publ)
+    await session.commit()
